requirements.txt
Selenium==4.21.0
pytest==8.2.2
behave==1.2.6
pytest-bdd==7.2.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
faker==24.9.0
